Swap out git commit command for a message
As written, this script commits the terraform state files to source
control and then asks the user to complete the process by pushing the
changes up.

However, if the user has any changes staged in source control when
running this script, they will be automatically committed along with the
terraform state files using our hard-coded commit message.

Also, the script will add this commit to whatever branch the user has
checked out when it is invoked.

This change removes the commands that affect git state, and extends the
manual intervention requested from the user to include staging &
committing the state files.